requirements.txt
kafka-python==2.0.2
minio==7.1.15
pyspark==3.5.1
boto3==1.34.84
faker==25.2.0
psycopg2-binary==2.9.9
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-apache-spark
